requirements.txt
# Core AI / LLM
anthropic>=0.40.0

# Local embedding model (no extra API key needed)
sentence-transformers>=3.0.0

# Used for fast cosine similarity search (replaces ChromaDB for Python 3.14 compatibility)
numpy>=1.26.0

# Web UI
streamlit>=1.40.0

# Web scraping for fetching Salesforce docs
beautifulsoup4>=4.12.0
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0
